exception

+++ app/Http/Controllers/FavoriteController.php
@@ -36,8 +36,13 @@ class FavoriteController extends Controller
         }
 
         $user = $this->userService->getByID(auth('api')->user()->id);
-        // $user = $this->userService->getByID(1);
-        $celebrity = $this->celebService->getModel(['external_id' => $params('celeb_ext_id')]);
+        $celebrity = $this->celebService->getModel(['external_id' => $params['celeb_ext_id']]);
+        if (!$celebrity) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Unregisterd celebrity'
+            ]);
+        }
         $user->favorites()->attach($celebrity, ['created_on' => Date('Y-m-d')]);
         return response()->json(['success' => true]);
     }
@@ -58,8 +63,13 @@ class FavoriteController extends Controller
         }
         
         $user = $this->userService->getByID(auth('api')->user()->id);
-        // $user = $this->userService->getByID(1);
-        $celebrity = $this->celebService->getModel(['external_id' => $params('celeb_ext_id')]);
+        $celebrity = $this->celebService->getModel(['external_id' => $params['celeb_ext_id']]);
+        if (!$celebrity) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Unregisterd celebrity'
+            ]);
+        }
         $user->favorites()->detach($celebrity);
         return response()->json(['success' => true]);
     }
@@ -70,7 +80,6 @@ class FavoriteController extends Controller
     public function show(Request $request)
     {   
         $user = $this->userService->getByID(auth('api')->user()->id);
-        // $user = $this->userService->getByID(1);
         $favorites = $user->favorites;
         $data = [];
         foreach ($favorites as $item) {
